movie data jsp파일에 적용 및 불필요한 코드 제거

+++ movie/src/main/java/jang/app/movie/reservations/ReservationEntity.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 public class ReservationEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int reservationId;
+    private Long reservationId;
 
     @Column(name = "reservation_data", nullable = false)
     private LocalDateTime reservationData;
