+++ movie/src/main/java/jang/app/movie/reservationseats/ReservationSeatId.java
@@ -12,12 +12,12 @@ import java.io.Serializable;
 public class ReservationSeatId implements Serializable {
 
     //복합키
-    private int showtimeId;
-    private int reservationId;
-    private int seatId;
+    private Integer showtimeId;
+    private Long reservationId;
+    private Integer seatId;
 
     @Builder
-    public ReservationSeatId(int showtimeId, int reservationId, int seatId) {
+    public ReservationSeatId(Integer showtimeId, Long reservationId, Integer seatId) {
         this.showtimeId = showtimeId;
         this.reservationId = reservationId;
         this.seatId = seatId;
